+++ SingleWord.cpp
@@ -25,7 +25,7 @@
 #include <cmath>
 #include <ctime>
 
-typedef unsigned short int ushort;
+typedef unsigned int uint;
 using namespace std;
 
 SingleWord::~SingleWord() {
@@ -36,7 +36,7 @@ SingleWord::SingleWord(const string &Aspelling, const string &Asound) {
 	q_meanings = 0;
 	known = false;
 	meanings = vector<SingleWord *>(0);
-	repetitionsOfMeanings = vector<ushort*>(0);
+	repetitionsOfMeanings = vector<uint*>(0);
 	lastRepetitionsOfM = vector<time_t*>(0);
 	flag_names = vector<string>(0);
 	flags = vector<string>(0);
@@ -47,11 +47,11 @@ SingleWord::SingleWord(SingleWord const *sw) {
 	q_meanings = sw->q_meanings;
 	meanings = vector<SingleWord*>(sw->meanings); // czy napewno skopiuje wektor
 	lastRepetitionsOfM = vector<time_t*>(sw->lastRepetitionsOfM);
-	repetitionsOfMeanings = vector<ushort*>(sw->repetitionsOfMeanings);
+	repetitionsOfMeanings = vector<uint*>(sw->repetitionsOfMeanings);
 	flag_names = sw->flag_names;
 	flags = sw->flags;
 }
-void SingleWord::setTimeLastRepetition(const ushort &number_meaning, const time_t &lasttime) {
+void SingleWord::setTimeLastRepetition(const uint &number_meaning, const time_t &lasttime) {
 	if(number_meaning >= q_meanings)throw Error::newError(Error::BAD_ARGUMENT, "", __LINE__, __FILE__);
 	
 	//veryfying correctness data
@@ -65,19 +65,19 @@ void SingleWord::setTimeLastRepetition(const ushort &number_meaning, const time_
 bool SingleWord::isKnown() const{
 	return known;
 }
-ushort SingleWord::getRepetitionLevelNumber(const ushort &number_meaning) const {
+uint SingleWord::getRepetitionLevelNumber(const uint &number_meaning) const {
 	if(number_meaning >= q_meanings)throw Error::newError(Error::BAD_ARGUMENT, "", __LINE__, __FILE__);
 	return *repetitionsOfMeanings[number_meaning];
 }
-void SingleWord::setRepetitionLevelNumber(const ushort &number_meaning, const ushort &which_repetition) const {
+void SingleWord::setRepetitionLevelNumber(const uint &number_meaning, const uint &which_repetition) const {
 	if(number_meaning >= q_meanings)throw Error::newError(Error::BAD_ARGUMENT, "", __LINE__, __FILE__);
 	*repetitionsOfMeanings[number_meaning] = which_repetition;
 }
-time_t SingleWord::getTimeLastRepetition(const ushort &number_meaning) const {
+time_t SingleWord::getTimeLastRepetition(const uint &number_meaning) const {
 	if(number_meaning >= q_meanings)throw Error::newError(Error::BAD_ARGUMENT, "", __LINE__, __FILE__);
 	return *lastRepetitionsOfM[number_meaning];
 }
-void SingleWord::connectSingleWords(SingleWord *sw1, SingleWord *sw2, const ushort &which_repetition, const ushort &which_repetition2, const time_t &last_repetition) {
+void SingleWord::connectSingleWords(SingleWord *sw1, SingleWord *sw2, const uint &which_repetition, const uint &which_repetition2, const time_t &last_repetition) {
 	//ASSERT IN
 	assert(sw1 != NULL && sw2 != NULL);
 	assert(!sw1->isConnectedWith(sw2));
@@ -85,7 +85,7 @@ void SingleWord::connectSingleWords(SingleWord *sw1, SingleWord *sw2, const usho
 	int temp1 = sw1->meanings.size();
 	sw1->meanings.push_back(sw2);
 	sw1->repetitionsOfMeanings.push_back(NULL);
-	sw1->repetitionsOfMeanings[temp1] = new ushort;
+	sw1->repetitionsOfMeanings[temp1] = new uint;
 	*(sw1->repetitionsOfMeanings[temp1]) = which_repetition;
 	sw1->lastRepetitionsOfM.push_back(NULL);
 
@@ -99,8 +99,8 @@ void SingleWord::connectSingleWords(SingleWord *sw1, SingleWord *sw2, const usho
 	int temp2 = sw2->meanings.size();
 	sw2->meanings.push_back(sw1);
 	sw2->repetitionsOfMeanings.push_back(NULL);
-	sw2->repetitionsOfMeanings[temp2] = new ushort;
-	*(sw2->repetitionsOfMeanings[temp2]) = (ushort)which_repetition2;//niewspólna cecha dla dwóch słów
+	sw2->repetitionsOfMeanings[temp2] = new uint;
+	*(sw2->repetitionsOfMeanings[temp2]) = (uint)which_repetition2;//niewspólna cecha dla dwóch słów
 	sw2->lastRepetitionsOfM.push_back(NULL); 
 	sw2->lastRepetitionsOfM[temp2] = sw1->lastRepetitionsOfM[temp1];//wspólna cecha dwa dwóch słów
 	sw2->q_meanings++;
@@ -166,19 +166,19 @@ bool SingleWord::isEmpty() const{
 	return q_meanings == 0 ? true : false;
 }
 bool SingleWord::isConnectedWith(SingleWord* meaning) const{
-	for(ushort i = 0; i < q_meanings; i++){
+	for(uint i = 0; i < q_meanings; i++){
 		if(meaning == meanings[i])return true;
 	}
 	return false;
 }
 void SingleWord::joinOtherSingleWord(SingleWord* sword) {
-	for(ushort i = 0; i < sword->q_meanings; i++) {
+	for(uint i = 0; i < sword->q_meanings; i++) {
 		if(isConnectedWith(sword->meanings[i]))continue;
 		else connectSingleWords(this, sword, 0, 0, 0);
 	}
 }
 int SingleWord::findMeaning(SingleWord *meaning) const{
-	for(ushort i = 0; i < q_meanings; i++){
+	for(uint i = 0; i < q_meanings; i++){
 		if(meaning == meanings[i])return i;
 	}
 	return -1;
@@ -195,11 +195,11 @@ void SingleWord::setSound(const string &Asound) {
 void SingleWord::setSpelling(const string &Aspelling) {
 	spelling = Aspelling;
 }
-SingleWord * SingleWord::getMeaning(const ushort &number) const {
+SingleWord * SingleWord::getMeaning(const uint &number) const {
 	if(number >= q_meanings)throw Error::newError(Error::BAD_ARGUMENT, "", __LINE__, __FILE__);
 	return meanings[number];
 }
-ushort SingleWord::getNumberMeanings() const {
+uint SingleWord::getNumberMeanings() const {
 	return q_meanings;
 }
 SingleWord SingleWord::newSingleWord(const string &spelling, const string &sound) {
